pyproject.toml
[project]
name = "mudaserver"
version = "0.1.0"
description = "REST API built with FastAPI for a home server hosted on Linux. "
authors = [
    {name = "Muhammad Moosa",email = "muhammadmoosa.7782@gmail.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "requests (>=2.32.3,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "py-cpuinfo (>=9.0.0,<10.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
]

[tool.poetry]
packages = [{include = "app"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "main:main"
